% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa2g.r
\name{calculate_aa2g}
\alias{calculate_aa2g}
\title{Calculate Age-Adjusted Incidence Rates}
\usage{
calculate_aa2g(in1, in2, level1, level2, verbose = FALSE)
}
\arguments{
\item{in1}{A data frame containing incidence data for the first group}

\item{in2}{A data frame containing incidence data for the second group}

\item{level1}{Name of the first group}

\item{level2}{Name of the second group}

\item{verbose}{Logical, if TRUE, print detailed output. Default is FALSE.}
}
\value{
A data frame with age-adjusted incidence rates
}
\description{
This function calculates age-adjusted incidence rates, similar to the AA2G macro in SAS.
}
