% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetime_risk.R
\name{calculate_cumulative_incidence}
\alias{calculate_cumulative_incidence}
\title{Calculate cumulative incidence with intermediate datasets}
\usage{
calculate_cumulative_incidence(data, age_free, debug = FALSE)
}
\arguments{
\item{data}{Person-year dataset}

\item{age_free}{Starting age for analysis}

\item{debug}{If TRUE, save intermediate datasets}
}
\value{
A list containing unadjusted and adjusted cumulative incidence
}
\description{
Calculate cumulative incidence with intermediate datasets
}
