% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incid.r
\name{calculate_incidence}
\alias{calculate_incidence}
\title{Calculate Incidence Rates}
\usage{
calculate_incidence(
  data,
  min_age,
  max_age,
  age_group_width,
  min_age_group,
  verbose = FALSE
)
}
\arguments{
\item{data}{A data frame containing the summary dataset}

\item{min_age}{Minimum age to consider}

\item{max_age}{Maximum age to consider}

\item{age_group_width}{Width of age groups}

\item{min_age_group}{Minimum age group to consider}

\item{verbose}{Logical, if TRUE, print detailed output. Default is FALSE.}
}
\value{
A data frame with incidence rates
}
\description{
This function calculates incidence rates, similar to the INCID macro in SAS.
}
