% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.r
\name{calculate_lr}
\alias{calculate_lr}
\title{Calculate Cumulative Incidence}
\usage{
calculate_lr(data, age_free, max_age, age_group_width, verbose = FALSE)
}
\arguments{
\item{data}{A data frame containing the input data}

\item{age_free}{Minimum age to consider}

\item{max_age}{Maximum age to consider}

\item{age_group_width}{Width of age groups}

\item{verbose}{Logical, if TRUE, print detailed output. Default is FALSE.}
}
\value{
A list containing cumulative incidence data and plots
}
\description{
This function calculates unadjusted cumulative incidence and cumulative incidence adjusted for competing risk,
similar to the LR macro in SAS.
}
