% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_analysis.R
\name{create_lifetime_risk_table}
\alias{create_lifetime_risk_table}
\title{Create Lifetime Risk Table from Different Starting Ages}
\usage{
create_lifetime_risk_table(
  data,
  index_ages,
  max_age,
  group_var,
  levels,
  adjusted = FALSE,
  age_group_width = 5
)
}
\arguments{
\item{data}{Input dataset}

\item{index_ages}{Vector of starting ages}

\item{max_age}{Maximum age to consider}

\item{group_var}{Group variable name}

\item{levels}{Vector of group levels to include}

\item{adjusted}{Logical, whether to use competing risk adjusted estimates}

\item{age_group_width}{Width of age groups (default = 5)}
}
\value{
A data frame containing lifetime risks from different starting ages
}
\description{
Create Lifetime Risk Table from Different Starting Ages
}
