% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetime_risk.R
\name{create_person_year_data}
\alias{create_person_year_data}
\title{Create person-year dataset for analysis}
\usage{
create_person_year_data(data, min_age, max_age, debug = FALSE)
}
\arguments{
\item{data}{Input dataset}

\item{min_age}{Minimum age to consider}

\item{max_age}{Maximum age to consider}

\item{debug}{If TRUE, save intermediate datasets}
}
\value{
A data.table with expanded person-year data
}
\description{
Create person-year dataset for analysis
}
