% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{create_summary_dataset}
\alias{create_summary_dataset}
\title{Create Summary Data Set for Each Age}
\usage{
create_summary_dataset(data, group, levels, verbose = FALSE)
}
\arguments{
\item{data}{A data frame containing the preprocessed data}

\item{group}{Group variable name}

\item{levels}{A vector of levels to include in the summary}

\item{verbose}{Logical, if TRUE, print detailed output. Default is FALSE.}
}
\value{
A data frame with summary statistics for each age
}
\description{
This function creates a summary data set for each age, similar to the SDSMAC macro in SAS.
}
