% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.r
\name{generate_lr_plot}
\alias{generate_lr_plot}
\title{Generate LR Plot}
\usage{
generate_lr_plot(data, title)
}
\arguments{
\item{data}{A data frame containing the LR calculation results}

\item{title}{Title for the plot}
}
\value{
A ggplot object
}
\description{
This function generates a plot of the LR calculation results.
}
