% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess Input Data}
\usage{
preprocess_data(data, max_age, min_age, group, verbose = FALSE)
}
\arguments{
\item{data}{A data frame containing the input data}

\item{max_age}{Maximum age to consider}

\item{min_age}{Minimum age to consider}

\item{group}{Group variable name}

\item{verbose}{Logical, if TRUE, print detailed output. Default is FALSE.}
}
\value{
A preprocessed data frame
}
\description{
This function preprocesses the input data according to the logic in the DATA1 macro.
}
