% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_analysis.R
\name{print.pie_analysis}
\alias{print.pie_analysis}
\title{Print method for pie_analysis objects}
\usage{
\method{print}{pie_analysis}(x, ...)
}
\arguments{
\item{x}{A pie_analysis object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for pie_analysis objects
}
