% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.r
\name{print.pie_lifetime_risk}
\alias{print.pie_lifetime_risk}
\title{Print method for pie_lifetime_risk objects}
\usage{
\method{print}{pie_lifetime_risk}(x, ...)
}
\arguments{
\item{x}{A pie_lifetime_risk object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for pie_lifetime_risk objects
}
