% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.r
\name{print_lr_results}
\alias{print_lr_results}
\title{Print LR Results}
\usage{
print_lr_results(data, age_free, age_group_width)
}
\arguments{
\item{data}{A data frame containing the LR calculation results}

\item{age_free}{Minimum age to consider}

\item{age_group_width}{Width of age groups}
}
\description{
This function prints the results of the LR calculation in a more readable format.
}
