% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample data}
\format{
\subsection{\code{sample_data}}{

A data frame with 1,000 rows and 6 columns:
\describe{
\item{ids}{Participant id}
\item{entryage, survage}{Age at entry and exit of the study}
\item{status}{Status at exit of the study (1 = death, 0 = censored)}
\item{astats}{?}
\item{group}{A grouping variable}
}
}
}
\source{
simulated
}
\usage{
sample_data
}
\description{
A simulated set of data used to demonstrate functions in \code{lifetimerisk}
}
\keyword{datasets}
