% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.r
\name{summarize_lifetime_risk}
\alias{summarize_lifetime_risk}
\title{Summarize Lifetime Risk}
\usage{
summarize_lifetime_risk(data)
}
\arguments{
\item{data}{A data frame containing the LR calculation results}
}
\value{
A string containing the summary
}
\description{
This function summarizes the key lifetime risk measurements.
}
