% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_analysis.R
\name{summary.pie_analysis}
\alias{summary.pie_analysis}
\title{Summary method for pie_analysis objects}
\usage{
\method{summary}{pie_analysis}(object, ...)
}
\arguments{
\item{object}{A pie_analysis object}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for pie_analysis objects
}
