% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.r
\name{summary.pie_lifetime_risk}
\alias{summary.pie_lifetime_risk}
\title{Summary method for pie_lifetime_risk objects}
\usage{
\method{summary}{pie_lifetime_risk}(object, ...)
}
\arguments{
\item{object}{A pie_lifetime_risk object}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for pie_lifetime_risk objects
}
