% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetime_risk.R
\name{create_summary_dataset}
\alias{create_summary_dataset}
\title{Create summary dataset (equivalent to SDSMAC)}
\usage{
create_summary_dataset(
  data,
  group_var,
  level,
  dataset_name = NULL,
  debug = FALSE
)
}
\arguments{
\item{data}{Person-year dataset}

\item{group_var}{Group variable name}

\item{level}{Group level(s)}

\item{dataset_name}{Name to save dataset as when debugging}

\item{debug}{If TRUE, save intermediate datasets}
}
\value{
A data.table with summary statistics
}
\description{
Create summary dataset (equivalent to SDSMAC)
}
